start separating out node/electron API usages
