+++ src/fileUtil.ts
@@ -1,0 +1,28 @@
+import {
+  openFile,
+  readFile,
+  writeFile,
+  decode,
+  decodeAttachment,
+  readFileSync,
+} from './fileUtilElectron';
+
+/*
+import {
+  openFile,
+  readFile,
+  writeFile,
+  decode,
+  decodeAttachment,
+  readFileSync,
+} from './fileUtilWeb';
+ */
+
+export {
+  openFile,
+  readFile,
+  writeFile,
+  decode,
+  decodeAttachment,
+  readFileSync,
+};
